chore: split BTAnalyais to metrics
