+++ freqtrade/plugins/protections/max_drawdown_protection.py
@@ -5,7 +5,7 @@ from typing import Any, Dict
 
 import pandas as pd
 
-from freqtrade.data.btanalysis import calculate_max_drawdown
+from freqtrade.data.metrics import calculate_max_drawdown
 from freqtrade.persistence import Trade
 from freqtrade.plugins.protections import IProtection, ProtectionReturn
